+++ packages/elysia/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vercel/elysia
 
+## 0.1.50
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.6.17
+
 ## 0.1.49
 
 ### Patch Changes